requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi== 0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pip==22.3.1
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
setuptools==65.6.3
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wheel==0.38.4